% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_trim.R
\name{bound_trim}
\alias{bound_trim}
\title{Bound trim the sequence vector}
\usage{
bound_trim(vector, min, max)
}
\arguments{
\item{vector}{대상이되는 sequence를 입력합니다. e.g \code{1:100}}

\item{min}{최소경계를 지정합니다. e.g \code{5}}

\item{max}{최대경계를 지정합니다. e.g \code{95}}
}
\description{
일반수열에서 최소치와 최대치를 넘어가는, 일종의 경계선을 넘겨버린 값 만을 제외시킨 결과를 반환합니다.
}
\examples{
bound_trim(1:100, 5, 95)
bound_trim(c(1, 5, 9, 10), 0, 9)
bound_trim(rnorm(100), -1.96, 1.96)
}

