% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoising.r
\name{denoising}
\alias{denoising}
\title{Denoising function}
\usage{
denoising(x, method = c("mean", "median", "MDMR", "cross_shape_mean",
  "cross_shape_median", "cross_shape_MDMR", "mean_oneframe", "median_oneframe",
  "auto_outlier_filter", "median_filter", "MDMR_filter"), W0 = 1, lambda,
  save.env = "bsid_env", output.naming = NULL, ...)
}
\arguments{
\item{x}{영상정보 matrix 를 입력합니다.}
}
\description{
Denoising function
}
\examples{

}

