% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.r
\name{outlier_detect}
\alias{outlier_detect}
\title{Outlier detect}
\usage{
outlier_detect(x)
}
\arguments{
\item{x}{numeric vector 을 입력합니다.}
}
\description{
Outlier detect
}
\examples{
set.seed(111)
r <- rnorm(100)
outlier_detect(r)
r[which(outlier_detect(r))]
boxplot(r)$out
}
