% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCP.R
\name{PCP}
\alias{PCP}
\title{PCP algorithm (미완성)}
\usage{
PCP(x, iter.max = 100, save.env = "bsid_env", output.naming = NULL)
}
\arguments{
\item{x}{영상정보 matrix 를 입력합니다.}

\item{save.env}{반환될 배경정보행렬 \code{L}, 전경정보행렬 \code{S} 객체의 환경을 지정합니다. 자세한 것은 value 장을 확인합니다.}
}
\value{
\code{save.env} envirnoment 에 배경정보행렬 \code{L}, 전경정보행렬 \code{S} 객체가 반환됩니다. 전역환경이 아닌 임시환경에 저장되는 이유는 덮어쓰기를 방지하기 위함입니다.
}
\description{
PCP algorithm
}
\details{
본 함수는 \code{Singular_value_thresholding_operator()}, \code{Shrinkage_operator()} 함수를 이용합니다.
}

