% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_vm}
\alias{plot_vm}
\title{Convert oneframe picture about video data matrix (Only black and white)}
\usage{
plot_vm(x, picture = 1, width)
}
\arguments{
\item{x}{영상정보 matrix 를 입력합니다.}

\item{picture}{변환을 원하는 프레임을 입력합니다. 정수만 허락 합니다.}

\item{nc}{영상 사진의 height 가 몇인지 입력합니다. 정수만 허락 합니다.}
}
\description{
영상정보 matrix 에 대하여 영상중 1프레임을 흑백사진으로 ploting 합니다.
}
\examples{
## not run
plot_vm(M, picture=1, nc=160)
}

