% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{save_anipic}
\alias{save_anipic}
\title{Convert animation GIF picture about video data matrix}
\usage{
save_anipic(x, interval = 0.1, save.name, ...)
}
\arguments{
\item{x}{영상정보 matrix 를 입력합니다.}

\item{interval}{animation 의 interval 을 설정합니다. 기본값은 0.1초 입니다.}

\item{save.name}{외부로 저장할 파일명을 문자열 형으로 설정합니다. 확장자는 \code{.gif} 로 이미 기 설정되어 있으므로 이를 고려하지 않습니다.}

\item{nc}{영상 사진의 height 가 몇인지 입력합니다. 정수만 허락 합니다.}
}
\value{
워킹디렉토리에 "\{save.name\}.gif" 파일이 저장됩니다.
}
\description{
영상정보 matrix 를 animationGIF 파일로 변환시킵니다.
}
\details{
본 함수는 \code{plot_vm()}, \code{animation::saveGIF()} 함수를 이용합니다.
따라서 animation package 종속입니다.
}
\examples{
## not run
save_anipic(M, nc=160)
}
\seealso{
\link{saveGIF}, \link{plot_vm}
}

