% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_trim.R
\name{window}
\alias{window}
\title{Bandwidth indexing with Bound trim}
\usage{
window(W0, c, min, max)
}
\arguments{
\item{W0}{Bandwidth 크기를 설정합니다.}

\item{c}{Bandwidth indexing 의 기준점을 설정합니다.}

\item{min}{최소경계를 지정합니다.}

\item{max}{최대경계를 지정합니다.}
}
\description{
Bandwidth indexing with Bound trim
}
\examples{
window(10, 0, -20, 20)
window(10, 0, -5, 20)
}

