% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhiteNoise.R
\name{WN_line}
\alias{WN_line}
\title{Line type white noise addition of vedio matrix}
\usage{
WN_line(x, line.N.weight = 0.1, line.sd = 10, line.length.weight = 0.3,
  save.env = "bsid_env", output.naming = NULL, direction = c("width",
  "height"))
}
\arguments{
\item{x}{영상정보 matrix 를 입력합니다.}

\item{save.env}{반환될 잡음추가 영상행렬 \code{Mstar} 객체의 환경을 지정합니다. 자세한 것은 value 장을 확인합니다.}
}
\value{
\code{save.env} envirnoment 에 잡음추가 영상행렬 \code{Mstar} 객체가 반환됩니다. 전역환경이 아닌 임시환경에 저장되는 이유는 덮어쓰기를 방지하기 위함입니다.
}
\description{
Line type white noise addition of vedio matrix.
$M$ matrix 에 white noise 를 부여하여 $M^\star$ 를 만들며, 잡음타입은 랜덤하게 부여된 line type white noise 입니다.
}
\examples{
WN_line(M, output.naming="Mstar")
}
