% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhiteNoise.R
\name{WN_point}
\alias{WN_point}
\title{Point type white noise addition of vedio matrix}
\usage{
WN_point(x, point.N.mean = attributes(x)$n1 * 0.01, point.N.sd = 10,
  save.env = "bsid_env", output.naming = NULL)
}
\arguments{
\item{x}{영상정보 matrix 를 입력합니다.}

\item{point.N.mean}{부여하고싶은 point white noise 개수의 평균을 입력합니다.}

\item{point.N.sd}{부여하고싶은 point white noise 개수의 표준편차를 입력합니다.}

\item{save.env}{반환될 잡음추가 영상행렬 \code{Mstar} 객체의 환경을 지정합니다. 자세한 것은 value 장을 확인합니다.}
}
\value{
\code{save.env} envirnoment 에 잡음추가 영상행렬 \code{Mstar} 객체가 반환됩니다. 전역환경이 아닌 임시환경에 저장되는 이유는 덮어쓰기를 방지하기 위함입니다.
}
\description{
Point type white noise addition of vedio matrix.
$M$ matrix 에 white noise 를 부여하여 $M^\star$ 를 만들며, 잡음타입은 랜덤하게 부여된 point type white noise 입니다.
}
\examples{
WN_point(M, output.naming="Mstar")
}

